requirements.txt
# requirements.txt
gspread
oauth2client
python-dotenv
google-api-python-client